backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.9
PyPDF2==3.0.1
langchain==0.1.5
faiss-cpu==1.7.4
numpy==1.26.3
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.3.2
sentence-transformers==2.5.1 